+++ src/main/java/jezzsantos/automate/plugin/application/AutomateApplication.java
@@ -9,6 +9,7 @@ import jezzsantos.automate.plugin.application.interfaces.EditingMode;
 import jezzsantos.automate.plugin.application.interfaces.drafts.DraftDetailed;
 import jezzsantos.automate.plugin.application.interfaces.drafts.DraftElement;
 import jezzsantos.automate.plugin.application.interfaces.drafts.DraftLite;
+import jezzsantos.automate.plugin.application.interfaces.drafts.LaunchPointExecutionResult;
 import jezzsantos.automate.plugin.application.interfaces.patterns.Attribute;
 import jezzsantos.automate.plugin.application.interfaces.patterns.PatternDetailed;
 import jezzsantos.automate.plugin.application.interfaces.patterns.PatternLite;
@@ -275,5 +276,12 @@ public class AutomateApplication implements IAutomateApplication {
     public void deleteDraftElement(@NotNull String expression) throws Exception {
 
         this.automateService.deleteDraftElement(expression);
+    }
+
+    @Override
+    @NotNull
+    public LaunchPointExecutionResult executeLaunchPoint(@NotNull String configurationPath, @NotNull String launchPointName) throws Exception {
+
+        return this.automateService.executeLaunchPoint(configurationPath, launchPointName);
     }
 }